Change float to double

+++ RayTracing/Models/Pixel.cs
@@ -1,10 +1,10 @@
 ﻿namespace RayTracing.Models {
 	class Pixel {
-		public float X { get; }
-		public float Y { get; }
+		public double X { get; }
+		public double Y { get; }
 		public Vector Color { get; }
 
-		public Pixel(float x, float y, Vector c) {
+		public Pixel(double x, double y, Vector c) {
 			X = x;
 			Y = y;
 			Color = c;
