+++ RayTracing/Primitives/IPrimitive.cs
@@ -5,6 +5,6 @@ namespace RayTracing.Primitives {
 	interface IPrimitive {
 		Color Color { get; }
 		int Specular { get; }
-		float Reflect { get; }
+		double Reflect { get; }
 	}
 }
