+++ RayTracing/Primitives/Light.cs
@@ -3,7 +3,7 @@
 namespace RayTracing.Primitives {
 	class Light {
 		public LightType Type { get; set; }
-		public float Intensity { get; set; }
+		public double Intensity { get; set; }
 		public Vector Position { get; set; }
 		public Vector Direction { get; set; }
 	}
